+++ rs/p2p/artifact_downloader/src/fetch_stripped_artifact/mod.rs
@@ -4,4 +4,7 @@ mod metrics;
 mod stripper;
 mod types;
 
+#[cfg(test)]
+mod test_utils;
+
 pub use assembler::FetchStrippedConsensusArtifact;
